refactor(cli): extract manifest skill apply domain
